+++ src/main/java/au/ellie/hyui/commands/HyUITestGuiCommand.java
@@ -160,7 +160,9 @@ public class HyUITestGuiCommand extends AbstractAsyncCommand {
                                     </span>
                                 </button>
                                 
-                               
+                               <div style="background-color: #2a2a3e; padding: 10; anchor-width: 140; align-items: center; justify-content: center;">
+                                        <img class="dynamic-image" src="https://hyvatar.io/render/Elyra" style="anchor-width: 120; anchor-height: 120;" />
+                               </div>
                                 <input type="number" value="42" min="-5" max="50" step="1.5" style="padding: 10; anchor-width: 50;"/>
                                 <p>Please enter your desired Buy It Now price:</p>
                                 <p style="color: #ff0000; visibility: hidden;" id="invalid-price">Price must be a positive number.</p>